requirements.txt
attr==0.3.2
cityscapesscripts==2.2.1
easydict==1.10
ftfy==6.0.3
imageio==2.9.0
ipython==8.11.0
matplotlib==3.1.2
mmcv==1.7.1
mmcv_full==1.3.0
numpy==1.16.4
opencv_python==3.4.2.16
packaging==20.4
Pillow==9.4.0
PyYAML==6.0
regex==2020.10.15
scikit_image==0.16.2
scikit_learn==1.2.2
scipy==1.2.1
skimage==0.0
tensorboardX==2.6
terminaltables==3.1.10
thop==0.1.1.post2209072238
timm==0.3.2
torch==1.8.1+cu101
torch_dct==0.1.5
torchsummary==1.5.1
torchvision==0.9.1+cu101
tqdm==4.51.0
yacs==0.1.8
